requirements.txt
appnope==0.1.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
decorator==4.4.0
defusedxml==0.5.0
entrypoints==0.3
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
google-api-core==1.14.2
google-auth==1.6.3
google-cloud==0.34.0
google-cloud-bigquery==1.20.0
google-cloud-core==1.0.3
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
gunicorn==19.9.0
idna==2.8
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jsonschema==3.0.2
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.4.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.2
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.2
psycopg2==2.8.3
ptyprocess==0.6.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
Pygments==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.2
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.22.0
rsa==4.0
Send2Trash==1.5.0
six==1.12.0
SQLAlchemy==1.3.8
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.2
urllib3==1.25.5
vaderSentiment==3.2.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
